+++ codex-rs/core/src/codex.rs
@@ -522,7 +522,7 @@ pub(crate) struct Session {
     /// session.
     features: Features,
     pending_mcp_server_refresh_config: Mutex<Option<McpServerRefreshConfig>>,
-    pub(crate) conversation: Arc<crate::conversation::RealtimeConversationManager>,
+    pub(crate) conversation: Arc<crate::realtime_conversation::RealtimeConversationManager>,
     pub(crate) active_turn: Mutex<Option<ActiveTurn>>,
     pub(crate) services: SessionServices,
     js_repl: Arc<JsReplHandle>,
@@ -1356,7 +1356,7 @@ impl Session {
             state: Mutex::new(state),
             features: config.features.clone(),
             pending_mcp_server_refresh_config: Mutex::new(None),
-            conversation: Arc::new(crate::conversation::RealtimeConversationManager::new()),
+            conversation: Arc::new(crate::realtime_conversation::RealtimeConversationManager::new()),
             active_turn: Mutex::new(None),
             services,
             js_repl,
@@ -7580,7 +7580,7 @@ mod tests {
             state: Mutex::new(state),
             features: config.features.clone(),
             pending_mcp_server_refresh_config: Mutex::new(None),
-            conversation: Arc::new(crate::conversation::RealtimeConversationManager::new()),
+            conversation: Arc::new(crate::realtime_conversation::RealtimeConversationManager::new()),
             active_turn: Mutex::new(None),
             services,
             js_repl,
@@ -7735,7 +7735,7 @@ mod tests {
             state: Mutex::new(state),
             features: config.features.clone(),
             pending_mcp_server_refresh_config: Mutex::new(None),
-            conversation: Arc::new(crate::conversation::RealtimeConversationManager::new()),
+            conversation: Arc::new(crate::realtime_conversation::RealtimeConversationManager::new()),
             active_turn: Mutex::new(None),
             services,
             js_repl,
