+++ codex-rs/core/src/lib.rs
@@ -13,7 +13,7 @@ pub mod auth;
 mod client;
 mod client_common;
 pub mod codex;
-mod conversation;
+mod realtime_conversation;
 pub use codex::SteerInputError;
 mod codex_thread;
 mod compact_remote;
